Refactor testimonial details
