+++ app/routes/o.$orgSlug/_public/testimonials/tmp.$id.tsx
@@ -1,5 +1,4 @@
 import { convexQuery } from "@convex-dev/react-query";
-import { useSuspenseQuery } from "@tanstack/react-query";
 import {
   createFileRoute,
   notFound,
@@ -8,34 +7,18 @@ import {
   useRouteContext,
   useRouter,
 } from "@tanstack/react-router";
-import { format } from "date-fns";
-import { ChevronLeft, TimerIcon } from "lucide-react";
+import { ChevronLeft } from "lucide-react";
 import { Suspense } from "react";
-import NotFound from "@/components/not-found";
-import TestimonialInfo from "@/components/testimonial-detail/testimonial-info";
-import TestimonialMedia from "@/components/testimonial-detail/testimonial-media";
-import TestimonialProcessingError from "@/components/testimonial-detail/testimonial-processing-error";
-import TestimonialSummary from "@/components/testimonial-detail/testimonial-summary";
-import TestimonialText from "@/components/testimonial-detail/testimonial-text";
-import { TestimonialTitle } from "@/components/testimonial-detail/testimonial-title";
 import { Button } from "@/components/ui/button";
-import {
-  Item,
-  ItemContent,
-  ItemDescription,
-  ItemMedia,
-  ItemTitle,
-} from "@/components/ui/item";
 import { Spinner } from "@/components/ui/spinner";
-import { TestimonialContext } from "@/contexts/testimonial-context";
 import { api } from "@/convex/_generated/api";
+import TempTestimonialDetail from "./-components/temp-testimonial-detail";
 
 export const Route = createFileRoute(
   "/o/$orgSlug/_public/testimonials/tmp/$id",
 )({
   ssr: false,
   component: Component,
-  notFoundComponent: NotFound,
   loader: async ({ context, params }) => {
     const { _id } = context.organization;
     const testimonial = await context.queryClient.ensureQueryData(
@@ -87,49 +70,5 @@ function Component() {
         <TempTestimonialDetail />
       </Suspense>
     </div>
-  );
-}
-
-function TempTestimonialDetail() {
-  const { id } = Route.useParams();
-  const { organization } = useRouteContext({ from: "/o/$orgSlug" });
-  const {
-    testimonial: preloadTestimonial,
-    expirationDate: preloadExpirationDate,
-  } = Route.useLoaderData();
-  const { data: liveTestimonial } = useSuspenseQuery(
-    convexQuery(api.testimonials.getTestimonialByIdAndOrgId, {
-      id: id,
-      orgId: organization._id,
-    }),
-  );
-  const testimonial = liveTestimonial || preloadTestimonial;
-  const liveExpirationDate = testimonial._creationTime + 900_000;
-  const expirationDate = liveExpirationDate || preloadExpirationDate;
-  return (
-    <TestimonialContext.Provider value={{ testimonial }}>
-      <div className="flex flex-col gap-8">
-        <Item variant="muted">
-          <ItemMedia variant="icon">
-            <TimerIcon />
-          </ItemMedia>
-          <ItemContent>
-            <ItemTitle>Recently Submitted Testimonial</ItemTitle>
-            <ItemDescription>
-              You can view this testimonial until{" "}
-              {format(expirationDate, "PPp")}
-            </ItemDescription>
-          </ItemContent>
-        </Item>
-        {testimonial.processingStatus === "error" && (
-          <TestimonialProcessingError />
-        )}
-        <TestimonialTitle />
-        {testimonial.media_type !== "text" && <TestimonialMedia />}
-        <TestimonialInfo />
-        <TestimonialSummary />
-        <TestimonialText />
-      </div>
-    </TestimonialContext.Provider>
   );
 }
